requirements.txt
PyPDF2
pytesseract
Pillow
pdf2image